Refactor enemy damage and health management: change damage and health types to int16_t for consistency and improved memory handling

+++ SOS/include/objects/enemy.h
@@ -35,9 +35,9 @@ public:
     void setTargetPlayer(std::shared_ptr<Player> player) { targetPlayer = player; }
 
     // Take damage from player attacks
-    virtual void takeDamage(int amount);
+    virtual void takeDamage(int16_t amount);
 
-    void setHealth(int newHealth);
+    void setHealth(int16_t newHealth);
 
     // Virtual methods to be overridden by derived classes
     virtual void move() = 0; // Move the enemy
